refactor(memory): single Desktop config-surface module
Memory providers already declare config (get_config_schema), persist it
(save_config), and read it (read_config). Desktop only needed a presentation
layer on top. Keep it to ONE module:

- hermes_cli/memory_provider_surface.py owns the whole Desktop surface:
  field kinds, tiers, label derivation, when-gating (field_visible),
  enrich_schema, and build_surface().
- The default read_config() is inlined on the MemoryProvider ABC (~20 lines,
  conventional <home>/<name>/config.json + env secrets), so agent/ stays
  self-contained and never imports the presentation layer.
- Honcho overrides read_config() for its host-keyed honcho.json.

No parallel registry, no per-provider UI code; all providers covered
generically. Presentation vocabulary stays out of the core runtime.
